894 asynkron opprettelse av notifications (#944)
* first version that should work

* refactoring

* Refactoring

* Function that works (tested)

* Add notification template validation in InitializeCorrespondencesHandler

* refactored with better parameters to hangfire

* small change

* Changed back to _ instead of :, refactored

* Removed entity object as parameter to hangfire

* rm using

---------
