+++ src/Altinn.Correspondence.Application/InitializeCorrespondences/InitializeCorrespondencesHandler.cs
@@ -1,40 +1,34 @@
 using Altinn.Correspondence.Application.CorrespondenceDueDate;
+using Altinn.Correspondence.Application.CreateNotification;
 using Altinn.Correspondence.Application.Helpers;
 using Altinn.Correspondence.Application.PublishCorrespondence;
 using Altinn.Correspondence.Common.Caching;
 using Altinn.Correspondence.Common.Helpers;
 using Altinn.Correspondence.Core.Models.Entities;
 using Altinn.Correspondence.Core.Models.Enums;
-using Altinn.Correspondence.Core.Models.Notifications;
 using Altinn.Correspondence.Core.Options;
 using Altinn.Correspondence.Core.Repositories;
 using Altinn.Correspondence.Core.Services;
 using Altinn.Correspondence.Core.Services.Enums;
 using Hangfire;
 using Microsoft.Extensions.Caching.Hybrid;
-using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using OneOf;
 using System.Security.Claims;
-using System.Text.Json;
 
 namespace Altinn.Correspondence.Application.InitializeCorrespondences;
 
 public class InitializeCorrespondencesHandler(
     InitializeCorrespondenceHelper initializeCorrespondenceHelper,
     IAltinnAuthorizationService altinnAuthorizationService,
-    IAltinnNotificationService altinnNotificationService,
     IAltinnRegisterService altinnRegisterService,
     ICorrespondenceRepository correspondenceRepository,
-    ICorrespondenceNotificationRepository correspondenceNotificationRepository,
     INotificationTemplateRepository notificationTemplateRepository,
-    ICorrespondenceStatusRepository correspondenceStatusRepository,
     IResourceRegistryService resourceRegistryService,
     IBackgroundJobClient backgroundJobClient,
     IDialogportenService dialogportenService,
     IContactReservationRegistryService contactReservationRegistryService,
-    IHostEnvironment hostEnvironment,
     IHybridCacheWrapper hybridCacheWrapper,
     HangfireScheduleHelper hangfireScheduleHelper,
     IOptions<GeneralSettings> generalSettings,
@@ -52,11 +46,11 @@ public class InitializeCorrespondencesHandler(
             }
         }
         var hasAccess = await altinnAuthorizationService.CheckAccessAsSender(
-            user,
-            request.Correspondence.ResourceId,
-            request.Correspondence.Sender.WithoutPrefix(),
-            null,
-            cancellationToken);
+        user,
+        request.Correspondence.ResourceId,
+        request.Correspondence.Sender.WithoutPrefix(),
+        null,
+        cancellationToken);
         if (!hasAccess)
         {
             return AuthorizationErrors.NoAccessToResource;
@@ -84,7 +78,7 @@ public class InitializeCorrespondencesHandler(
             return CorrespondenceErrors.DueDateRequired;
         }
         var contactReservation = await HandleContactReservation(request);
-        if(contactReservation.TryPickT1(out var error, out var reservedRecipients))
+        if (contactReservation.TryPickT1(out var error, out var reservedRecipients))
         {
             return error;
         }
@@ -137,16 +131,12 @@ public class InitializeCorrespondencesHandler(
         {
             attachmentsToBeUploaded.AddRange(existingAttachments.Where(a => a != null).Select(a => a!));
         }
-        List<NotificationContent>? notificationContents = null;
+
+        // Validate notification content if notification is provided
         if (request.Notification != null)
         {
             var templates = await notificationTemplateRepository.GetNotificationTemplates(request.Notification.NotificationTemplate, cancellationToken, request.Correspondence.Content?.Language);
             if (templates.Count == 0)
-            {
-                return NotificationErrors.TemplateNotFound;
-            }
-            notificationContents = await GetNotificationContent(request.Notification, templates, request.Correspondence, cancellationToken, request.Correspondence.Content?.Language);
-            if (notificationContents.Count == 0)
             {
                 return NotificationErrors.TemplateNotFound;
             }
@@ -163,9 +153,10 @@ public class InitializeCorrespondencesHandler(
             return uploadError;
         }
 
+
         return await TransactionWithRetriesPolicy.Execute(async (cancellationToken) =>
         {
-            return await InitializeCorrespondences(request, attachmentsToBeUploaded, notificationContents, partyUuid, reservedRecipients, cancellationToken);
+            return await InitializeCorrespondences(request, attachmentsToBeUploaded, partyUuid, reservedRecipients, cancellationToken);
         }, logger, cancellationToken);
     }
 
@@ -193,7 +184,7 @@ public class InitializeCorrespondencesHandler(
             throw;
         }
     }
-    private async Task<OneOf<InitializeCorrespondencesResponse, Error>> InitializeCorrespondences(InitializeCorrespondencesRequest request, List<AttachmentEntity> attachmentsToBeUploaded, List<NotificationContent>? notificationContents, Guid partyUuid, List<string> reservedRecipients, CancellationToken cancellationToken)
+    private async Task<OneOf<InitializeCorrespondencesResponse, Error>> InitializeCorrespondences(InitializeCorrespondencesRequest request, List<AttachmentEntity> attachmentsToBeUploaded, Guid partyUuid, List<string> reservedRecipients, CancellationToken cancellationToken)
     {
         logger.LogInformation("Initializing {correspondenceCount} correspondences for {resourceId}", request.Recipients.Count, request.Correspondence.ResourceId);
         var correspondences = new List<CorrespondenceEntity>();
@@ -228,14 +219,15 @@ public class InitializeCorrespondencesHandler(
         {
             logger.LogInformation("Correspondence {correspondenceId} initialized", correspondence.Id);
             var dialogJob = backgroundJobClient.Enqueue(() => CreateDialogportenDialog(correspondence.Id));
-            await hybridCacheWrapper.SetAsync("dialogJobId_" + correspondence.Id, dialogJob, new HybridCacheEntryOptions
+            await hybridCacheWrapper.SetAsync($"dialogJobId_{correspondence.Id}", dialogJob, new HybridCacheEntryOptions
             {
                 Expiration = TimeSpan.FromHours(24)
             });
-            if (request.Correspondence.Content.Attachments.Count == 0) 
+            if (request.Correspondence.Content.Attachments.Count == 0)
             {
-                await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence, cancellationToken);
+                await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence.Id, cancellationToken);
             }
+
             var isReserved = correspondence.GetHighestStatus()?.Status == CorrespondenceStatus.Reserved;
             var notificationDetails = new List<InitializedCorrespondencesNotifications>();
             if (!isReserved)
@@ -248,202 +240,33 @@ public class InitializeCorrespondencesHandler(
 
                 if (request.Notification != null)
                 {
-                    var notifications = await CreateNotifications(request.Notification, correspondence, notificationContents, cancellationToken);
-                    foreach (var notification in notifications)
-                    {
-                        var notificationOrder = await altinnNotificationService.CreateNotification(notification, cancellationToken);
-                        if (notificationOrder is null)
-                        {
-                            notificationDetails.Add(new InitializedCorrespondencesNotifications()
-                            {
-                                OrderId = Guid.Empty,
-                                Status = InitializedNotificationStatus.Failure
-                            });
-                            continue;
-                        }
-                        var entity = new CorrespondenceNotificationEntity()
-                        {
-                            Created = DateTimeOffset.UtcNow,
-                            NotificationChannel = request.Notification.NotificationChannel,
-                            NotificationTemplate = request.Notification.NotificationTemplate,
-                            CorrespondenceId = correspondence.Id,
-                            NotificationOrderId = notificationOrder.OrderId,
-                            RequestedSendTime = notification.RequestedSendTime,
-                            IsReminder = notification.RequestedSendTime != notifications[0].RequestedSendTime,
-                            OrderRequest = JsonSerializer.Serialize(notification)
-                        };
-
-                        notificationDetails.Add(new InitializedCorrespondencesNotifications()
-                        {
-                            OrderId = entity.NotificationOrderId,
-                            IsReminder = entity.IsReminder,
-                            // For custom recipients, RecipientLookup will be null. As such, this also maps to Success
-                            Status = notificationOrder.RecipientLookup?.Status == RecipientLookupStatus.Failed ? InitializedNotificationStatus.MissingContact : InitializedNotificationStatus.Success
-                        });
-                        await correspondenceNotificationRepository.AddNotification(entity, cancellationToken);
-                        backgroundJobClient.ContinueJobWith<IDialogportenService>(dialogJob, (dialogportenService) => dialogportenService.CreateInformationActivity(correspondence.Id, DialogportenActorType.ServiceOwner, DialogportenTextType.NotificationOrderCreated, notification.RequestedSendTime.ToString("yyyy-MM-dd HH:mm")), JobContinuationOptions.OnlyOnSucceededState);
-                    }
+                    // Schedule notification creation as a background job
+                    var notificationJob = backgroundJobClient.Enqueue<CreateNotificationHandler>((handler) => handler.Process(new CreateNotificationRequest 
+                    { 
+                        NotificationRequest = request.Notification, 
+                        CorrespondenceId = correspondence.Id,
+                        Language = correspondence.Content != null ? correspondence.Content.Language : null,
+                        // RequestCorrespondence = correspondence
+                    }, cancellationToken));
                 }
             }
             if (request.Correspondence.Content.Attachments.Count > 0 && await correspondenceRepository.AreAllAttachmentsPublished(correspondence.Id, cancellationToken))
             {
-                await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence, cancellationToken);
+                await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence.Id, cancellationToken);
             }
             initializedCorrespondences.Add(new InitializedCorrespondences()
             {
                 CorrespondenceId = correspondence.Id,
                 Status = correspondence.GetHighestStatus().Status,
-                Recipient = correspondence.Recipient,
-                Notifications = notificationDetails
+                Recipient = correspondence.Recipient
             });
         }
+
         return new InitializeCorrespondencesResponse()
         {
             Correspondences = initializedCorrespondences,
             AttachmentIds = correspondences.SelectMany(c => c.Content?.Attachments.Select(a => a.AttachmentId)).Distinct().ToList()
         };
-    }
-
-    private async Task<List<NotificationOrderRequest>> CreateNotifications(NotificationRequest notification, CorrespondenceEntity correspondence, List<NotificationContent> contents, CancellationToken cancellationToken)
-    {
-        var notifications = new List<NotificationOrderRequest>();
-        string recipientWithoutPrefix = correspondence.Recipient.WithoutPrefix();
-        bool isOrganization = recipientWithoutPrefix.IsOrganizationNumber();
-        bool isPerson = recipientWithoutPrefix.IsSocialSecurityNumber();
-
-        var recipientOverrides = notification.CustomNotificationRecipients ?? [];
-        var newRecipients = new List<Recipient>();
-        foreach (var recipientOverride in recipientOverrides)
-        {
-            newRecipients.AddRange(recipientOverride.Recipients.Select(r => new Recipient
-            {
-                EmailAddress = r.EmailAddress,
-                MobileNumber = r.MobileNumber,
-                IsReserved = r.IsReserved,
-                OrganizationNumber = r.OrganizationNumber,
-                NationalIdentityNumber = r.NationalIdentityNumber
-            }));
-        }
-
-        List<Recipient> relevantRecipients = newRecipients.Count > 0 ? newRecipients : new List<Recipient>
-        {
-            new()
-            {
-                OrganizationNumber = isOrganization ? recipientWithoutPrefix : null,
-                NationalIdentityNumber = isPerson ? recipientWithoutPrefix : null
-            }
-        };
-
-        NotificationContent? content = null;
-        if (isOrganization)
-        {
-            content = contents.FirstOrDefault(c => c.RecipientType == RecipientType.Organization) ?? contents.FirstOrDefault(c => c.RecipientType == null);
-        }
-        else if (isPerson)
-        {
-            content = contents.FirstOrDefault(c => c.RecipientType == RecipientType.Person) ?? contents.FirstOrDefault(c => c.RecipientType == null);
-        }
-        await SetRecipientNameOnNotificationContent(content, correspondence.Recipient, cancellationToken);
-        var notificationOrder = new NotificationOrderRequest
-        {
-            IgnoreReservation = correspondence.IgnoreReservation,
-            Recipients = relevantRecipients,
-            ResourceId = correspondence.ResourceId,
-            RequestedSendTime = correspondence.RequestedPublishTime.UtcDateTime <= DateTime.UtcNow ? DateTime.UtcNow.AddMinutes(5) : correspondence.RequestedPublishTime.UtcDateTime.AddMinutes(5),
-            SendersReference = correspondence.SendersReference,
-            ConditionEndpoint = CreateConditionEndpoint(correspondence.Id.ToString()),
-            NotificationChannel = notification.NotificationChannel,
-            EmailTemplate = !string.IsNullOrWhiteSpace(content.EmailSubject) && !string.IsNullOrWhiteSpace(content.EmailBody) ? new EmailTemplate
-            {
-                Subject = content.EmailSubject,
-                Body = content.EmailBody,
-            } : null,
-            SmsTemplate = !string.IsNullOrWhiteSpace(content.SmsBody) ? new SmsTemplate
-            {
-                Body = content.SmsBody,
-            } : null
-        };
-        notifications.Add(notificationOrder);
-        if (notification.SendReminder)
-        {
-            notifications.Add(new NotificationOrderRequest
-            {
-                IgnoreReservation = correspondence.IgnoreReservation,
-                Recipients = relevantRecipients,
-                ResourceId = correspondence.ResourceId,
-                RequestedSendTime = hostEnvironment.IsProduction() ? notificationOrder.RequestedSendTime.AddDays(7) : notificationOrder.RequestedSendTime.AddHours(1),
-                ConditionEndpoint = CreateConditionEndpoint(correspondence.Id.ToString()),
-                SendersReference = correspondence.SendersReference,
-                NotificationChannel = notification.ReminderNotificationChannel ?? notification.NotificationChannel,
-                EmailTemplate = !string.IsNullOrWhiteSpace(content.ReminderEmailSubject) && !string.IsNullOrWhiteSpace(content.ReminderEmailBody) ? new EmailTemplate
-                {
-                    Subject = content.ReminderEmailSubject,
-                    Body = content.ReminderEmailBody,
-                } : null,
-                SmsTemplate = !string.IsNullOrWhiteSpace(content.ReminderSmsBody) ? new SmsTemplate
-                {
-                    Body = content.ReminderSmsBody,
-                } : null
-            });
-        }
-        return notifications;
-    }
-    private async Task SetRecipientNameOnNotificationContent(NotificationContent? content, string recipient, CancellationToken cancellationToken)
-    {
-        if (content == null)
-        {
-            return;
-        }
-        var recipientName = await altinnRegisterService.LookUpName(recipient.WithoutPrefix(), cancellationToken);
-        if (string.IsNullOrEmpty(recipientName))
-        {
-            return;
-        }
-        content.EmailBody = content.EmailBody?.Replace("$correspondenceRecipientName$", recipientName);
-        content.EmailSubject = content.EmailSubject?.Replace("$correspondenceRecipientName$", recipientName);
-        content.SmsBody = content.SmsBody?.Replace("$correspondenceRecipientName$", recipientName);
-        content.ReminderEmailBody = content.ReminderEmailBody?.Replace("$correspondenceRecipientName$", recipientName);
-        content.ReminderEmailSubject = content.ReminderEmailSubject?.Replace("$correspondenceRecipientName$", recipientName);
-        content.ReminderSmsBody = content.ReminderSmsBody?.Replace("$correspondenceRecipientName$", recipientName);
-    }
-    private async Task<List<NotificationContent>> GetNotificationContent(NotificationRequest request, List<NotificationTemplateEntity> templates, CorrespondenceEntity correspondence, CancellationToken cancellationToken, string? language = null)
-    {
-        var content = new List<NotificationContent>();
-        var sendersName = correspondence.MessageSender;
-        if (string.IsNullOrEmpty(sendersName))
-        {
-            sendersName = await altinnRegisterService.LookUpName(correspondence.Sender.WithoutPrefix(), cancellationToken);
-        }
-        foreach (var template in templates)
-        {
-            content.Add(new NotificationContent()
-            {
-                EmailSubject = CreateNotificationContentFromToken(template.EmailSubject, request.EmailSubject).Replace("$sendersName$", sendersName),
-                EmailBody = CreateNotificationContentFromToken(template.EmailBody, request.EmailBody).Replace("$sendersName$", sendersName),
-                SmsBody = CreateNotificationContentFromToken(template.SmsBody, request.SmsBody).Replace("$sendersName$", sendersName),
-                ReminderEmailBody = CreateNotificationContentFromToken(template.ReminderEmailBody, request.ReminderEmailBody).Replace("$sendersName$", sendersName),
-                ReminderEmailSubject = CreateNotificationContentFromToken(template.ReminderEmailSubject, request.ReminderEmailSubject).Replace("$sendersName$", sendersName),
-                ReminderSmsBody = CreateNotificationContentFromToken(template.ReminderSmsBody, request.ReminderSmsBody).Replace("$sendersName$", sendersName),
-                Language = template.Language,
-                RecipientType = template.RecipientType
-            });
-        }
-        return content;
-    }
-
-    private Uri? CreateConditionEndpoint(string correspondenceId)
-    {
-        var conditionEndpoint = new Uri($"{_generalSettings.CorrespondenceBaseUrl.TrimEnd('/')}/correspondence/api/v1/correspondence/{correspondenceId}/notification/check");
-        if (conditionEndpoint.Host == "localhost")
-        {
-            return null;
-        }
-        return conditionEndpoint;
-    }
-
-    private string CreateNotificationContentFromToken(string message, string? token = "")
-    {
-        return message.Replace("{textToken}", token + " ").Trim();
     }
 
     public async Task CreateDialogportenDialog(Guid correspondenceId)
@@ -452,15 +275,8 @@ public class InitializeCorrespondencesHandler(
         await correspondenceRepository.AddExternalReference(correspondenceId, ReferenceType.DialogportenDialogId, dialogId);
     }
 
-    internal class NotificationContent
+    public void SchedulePublish(Guid correspondenceId, DateTimeOffset publishTime, CancellationToken cancellationToken)
     {
-        public string? EmailSubject { get; set; }
-        public string? EmailBody { get; set; }
-        public string? SmsBody { get; set; }
-        public string? ReminderEmailBody { get; set; }
-        public string? ReminderEmailSubject { get; set; }
-        public string? ReminderSmsBody { get; set; }
-        public string? Language { get; set; }
-        public RecipientType? RecipientType { get; set; }
+        backgroundJobClient.Schedule<PublishCorrespondenceHandler>((handler) => handler.Process(correspondenceId, null, cancellationToken), publishTime);
     }
 }
