+++ src/Altinn.Correspondence.Application/MalwareScanResult/MalwareScanResultHandler.cs
@@ -116,7 +116,7 @@ public class MalwareScanResultHandler(
                     PartyUuid = partyUuid
                 }
             );
-            await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence, cancellationToken);
+            await hangfireScheduleHelper.SchedulePublishAfterDialogCreated(correspondence.Id, cancellationToken);
         }
         await correspondenceStatusRepository.AddCorrespondenceStatuses(list, cancellationToken);
         return;
