+++ src/Altinn.Correspondence.Core/Services/Enums/AltinnEventType.cs
@@ -21,4 +21,6 @@ public enum AltinnEventType
     CorrespondenceReceiverNeverConfirmed,
     CorrespondenceReceiverReserved,
     CorrespondenceReceiverNeverRead,
+    NotificationCreated,
+    CorrespondenceNotificationCreationFailed,
 }
